Adding support for setting User-Agent and other openURI options.
The one thing missing in other Vimeo/Youtube related gems is lack of
option to set own "User-Agent" header. Vimeo recommends setting it as
some libraries default headers are banned (and maybe the list will be
bigger in the future).

Options are well described in openURI docs. There was no reason to write
specs for it as it's just a kind of proxy to openURI options - and openURI
has it tested for sure.